pyproject.toml
[tool.poetry]
name = "mlbusters"
version = "0.1.0"
description = ""
authors = ["Antony Redman <RumataYounger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
langchain = "^0.2.2"
pydantic = "^2.7.3"
peft = "^0.11.1"
torch = "^2.3.1"
transformers = "^4.41.2"
uvicorn = "^0.30.1"
pillow = "^10.3.0"
scenedetect = "^0.6.3"
faiss-cpu = "^1.8.0"
pymongo = "^4.7.3"
faststream = "^0.5.11"
opencv-python = "^4.10.0.82"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
